removed for loop from border function
